// TheanUI/src/app/Model/DeliveryLocationDTO.ts
export class DeliveryLocationAPIDTO {
  userGuid: any;
  lat: any;
  logt: any;
  addr: any;
  junName: any;
  nickName: any;
}

export class DeliveryLocationDTO {
  userGuid: any;
  lat: any;
  logt: any;
  junName: any;
  nickName: any;
  houseName: any;
  houseNo: any;
  street: any;
  place: any;
}